Better integration of Bi-AGS
